+++ src/main/java/de/samply/teiler/TeilerConst.java
@@ -9,6 +9,7 @@ public class TeilerConst {
     public final static String CONVERTER_TEMPLATE_DIRECTORY = "CONVERTER_TEMPLATE_DIRECTORY";
     public final static String WRITE_FILE_DIRECTORY = "WRITE_FILE_DIRECTORY";
     public final static String CONVERTER_PATH = "CONVERTER_PATH";
+    public final static String CONVERTER_XML_APPLICATION_CONTEXT_PATH = "CONVERTER_XML_APPLICATION_CONTEXT_PATH";
 
     // Spring Values (SV)
     public final static String HEAD_SV = "${";
@@ -17,6 +18,7 @@ public class TeilerConst {
         + ":#{'./templates'}" + BOTTOM_SV;
     public final static String WRITE_FILE_DIRECTORY_SV = HEAD_SV + WRITE_FILE_DIRECTORY + ":#{'./output'}" + BOTTOM_SV;
     public final static String CONVERTER_PATH_SV = HEAD_SV + CONVERTER_PATH + ":#{'./converter/converter.xml'}" + BOTTOM_SV;
+    public final static String CONVERTER_XML_APPLICATION_CONTEXT_PATH_SV = HEAD_SV + CONVERTER_XML_APPLICATION_CONTEXT_PATH + ":#{'./converter/converter.xml'}" + BOTTOM_SV;
 
     // REST Paths
     public static final String REQUEST = "request";